Update botocore client and use cloned novaclient for functional tests
Since botocore 1.0.0 is released we make ec2api functional tests
to be compatible with it.

Previously used novaclient modification (v14 of
https://review.openstack.org/#/c/152569/) is too obsolete. It is not
compatible with modern oslo libraries. Current version of this
modification (v34) is obsolete as well.

The proposed solution is to use a separate novaclient repository, which
can be maintained independently of community. Until microversions are
not supported by novaclient upstream.

Change-Id: I75e1bbcd81678a56a0940b696cfa5dbc2c159cc6