Fix creation of EBS images in functional tests
OpenStack cannot delete a volume leaving its snapshots. As a result
volumes of an instance which is base of an created EBS image are not
deleted. This leads to quota faults in tests launched later.

This patch add volumes of EBS instance to cleanup storag to delete them
automatically at test's teardown operation.

Change-Id: Ie34d26d4f10a2a4b5ebeff93a06d8fd43e305fc1