Introduce a function to parse image device mappings
Images have two kinds of device mappings attribute: virtual (mappings)
and block (block_device_mapping). The last one may be stored in two
different formats (legacy and bdm v2). Device descriptions are merged
from this two attributes to a single list used for instance boot
operation.

We should have a function which parses and merges image device mappings
and returns the result device list of a certain format to use it in
various parts of ec2api.

This patch introduce the function which implements the requirements
above and returns device mappings in bdm v2 format.

Change-Id: I7ad8673dcb6ddba11c6e00e22aa68a35ea1ec0fd