Make _parse_block_device_mapping to return only passed attributes
Currently this function set default values for certain bdm attributes,
such as delete_on_termination, or boot_index. Also it returns mappings
wich is not fully bdm v2 compatible. This prevents to use it for wide
range of purposes.

This patch make the function to return fully bdm v2 compatible mapping
with only passed (and unambiguously calculated) attributes are set.

Change-Id: I4a5ec244cc1d41855128090127a3b0fa266891c0