Fix the first describing of an alien image by id
Currently if an image has not even been described, but a user gets its
id (from describe instance results for example), describe image with
selective filter operation will fail with InvalidAMIIDNotFound
exception.

The reason is that the mapping db object (id-os_id pair) is created, but
has no project_id set. Project_id was unknown when the mapping was being
created. So both requests to db by current project and for public
objects do not get this mapping.

Change-Id: I80dacdf4d3d7c317c3e96a6f0d45923e41cc5a40