Fix names of new devices in functional tests
Some tests fail in gating, since that Nova assigns true device names to
bdm records. These tests try to add a volume to an instance on 'd' or
'h' letter, when the only 'a' volume exists in the instance. But libvirt
reassigns the device name to avoid a hole in device name sequence. So
the name of the new device becomes 'b'.

This patch uses 'b' for new devices to workaround the feature.

Depends-On: I770d394f65e5be6a5f8f36c673806f91509e0abe
Change-Id: Iea8bb3fb03aead5c347d129f14b2abd1d2bb4626