Allow to set compute service type
Nova provides several API versions. Each version is provided on its own
endpoint. Endpoints are registered in Keystone service catalog with
different service types. Only one 'compute' service type name is
reserved for compute services. So names of other compute service types
are depended of custom deployment.

ec2-api prefers to work against v2.1 Nova API which allows to maximize
AWS compatibility. But service type provides v2.1 is unknown. It may be
not 'compute' one, as it is in DevStack. So ec2-api should allow to set
service type name to use it.

This patch adds an ability for service type to be specified. Old manner
of verification that used service type supports required API version is
inapporpriate now. So this check it's dropped, but a new one will be
added in a separate changed.

Change-Id: I59da2d7dada07d2f1b8e25ca17161b469a00fdfe