Honour API_WORKERS setting in devstack plugin
DevStack has API_WORKERS variable which mean a count of worker threads
for services. This variable allows to be set in local.conf. Otherwise it
is evaluated in stackrc script. DevStack's nova script uses it to set
workers count for ec2 and metadata services. Thus ec2-api have to do the
same.

Change-Id: I67a991be239588beb8086b7c462e559cd08da055