Create an image in a separate thread
Creating of an image by a running instance stops the instance. This
takes a time, and a client can resend the request due to timeout. The
resended request fails because instance is already not in appropriate
state (it's stoping, stopped, or scheduling at the moment).

Thus ec2api should make such images asynchronously. It should create a
DB item to reserve the image id, return it to a caller, and then stop
the instance and do other things. While image is asynchronously created
and no Glance image exists, ec2api should return the image in responses
for describe operation. Describe operation should return the image with
and without selective filter by the image id passed in.

Also it should consider a race when asynchronous task has not finished,
but Glance image is already created, and another client describe images.
In that case ec2api should not create a new db item, but reuse existed
one.

If the creation failed, ec2api should expose the image as failed and
allow to delete it.

Unit tests for auto_update_db and handle_unpaired_item will be added in
a separate review, because some refactoring is required for that.

Change-Id: Iba4ade5000bea4af35d97abd1e708bbbc57f372a