Use a singleton for admin session
It's not so efficiently to keep admin context per green thread, because
there is a lot of such threads (1k per cpu by default). This doesn't
save more requests to keystone for admin (re)authentification.

There are different strategies to decrease these requests, but the
simplest one is selected - singleton. It bases on an assumption that
Keystone's session is thread safe. It seems to be true.

Change-Id: I626815df693934825e281a834d9b1de852afcdf5