Switch to Nova v2.32 API
At v2.33 new bdm field 'tag' was introduced. This field is automatically
created by snapshot operation for volume backed instance (ebs). The
result of the operation - bdm - is stored in metadata of the snapshot
image. When ec2api user requires to run an instance with this image
changing some attributes of the attached volume (e.g.
delete_on_termination), this bdm is merged with user defined parameters
and sent to Nova. As the result the new 'tag' field is sent to Nova
since v2.32.

However the used API version is v2.10, which does not support 'tag'
field. This did not lead to faults (until I56348dc2) because Nova
verified the first bdm only, but the first bdm for ebs instances was a
bdm for the image. With the noted fix this no more works.

This patch swithes ec2api to use v2.32 to get 'tag' field supported.

Change-Id: I68329bbffeeff5d460f3fca1a212ba20b35fc284
(cherry picked from commit 480dc02de0d8413aa518a23b22a0140013df1350)