fix ec2-api-metadata memcached overlap with nova-api
When ec2-api-metadata and nova-api are using the same
memcached server as a cache backend, the cached values
get intermixed due to the fact that the cache keys are
the same. The result is that either nova-api or
ec2-api-metadata will report cache value errors.

This commit fixes this potential problem by using
a different cache key prefix for the ec2-api metadata
entries.

Change-Id: Ibd26b33b3370d5c9c6cca0bb6e0d9c9eb03c0d50
Closes-Bug: #1717494